Move Level to its own file.
