+++ logger.go
@@ -9,23 +9,7 @@ import (
 	"sort"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
-)
-
-// Level holds a severity level.
-type Level uint32
-
-// The severity levels. Higher values are more considered more
-// important.
-const (
-	UNSPECIFIED Level = iota
-	TRACE
-	DEBUG
-	INFO
-	WARNING
-	ERROR
-	CRITICAL
 )
 
 // A Logger represents a logging module. It has an associated logging
@@ -53,37 +37,6 @@ var (
 		"": root,
 	}
 )
-
-func (level Level) String() string {
-	switch level {
-	case UNSPECIFIED:
-		return "UNSPECIFIED"
-	case TRACE:
-		return "TRACE"
-	case DEBUG:
-		return "DEBUG"
-	case INFO:
-		return "INFO"
-	case WARNING:
-		return "WARNING"
-	case ERROR:
-		return "ERROR"
-	case CRITICAL:
-		return "CRITICAL"
-	}
-	return "<unknown>"
-}
-
-// get atomically gets the value of the given level.
-func (level *Level) get() Level {
-	return Level(atomic.LoadUint32((*uint32)(level)))
-}
-
-// set atomically sets the value of the receiver
-// to the given level.
-func (level *Level) set(newLevel Level) {
-	atomic.StoreUint32((*uint32)(level), uint32(newLevel))
-}
 
 // getLoggerInternal assumes that the modulesMutex is locked.
 func getLoggerInternal(name string) Logger {
@@ -215,29 +168,6 @@ func ResetLoggers() {
 			module.level.set(UNSPECIFIED)
 		}
 	}
-}
-
-// ParseLevel converts a string representation of a logging level to a
-// Level. It returns the level and whether it was valid or not.
-func ParseLevel(level string) (Level, bool) {
-	level = strings.ToUpper(level)
-	switch level {
-	case "UNSPECIFIED":
-		return UNSPECIFIED, true
-	case "TRACE":
-		return TRACE, true
-	case "DEBUG":
-		return DEBUG, true
-	case "INFO":
-		return INFO, true
-	case "WARN", "WARNING":
-		return WARNING, true
-	case "ERROR":
-		return ERROR, true
-	case "CRITICAL":
-		return CRITICAL, true
-	}
-	return UNSPECIFIED, false
 }
 
 func (logger Logger) getModule() *module {
