+++ logger_test.go
@@ -154,88 +154,6 @@ func (*loggerSuite) TestLevelsInherited(c *gc.C) {
 	c.Assert(second.EffectiveLogLevel(), gc.Equals, loggo.INFO)
 }
 
-var parseLevelTests = []struct {
-	str   string
-	level loggo.Level
-	fail  bool
-}{{
-	str:   "trace",
-	level: loggo.TRACE,
-}, {
-	str:   "TrAce",
-	level: loggo.TRACE,
-}, {
-	str:   "TRACE",
-	level: loggo.TRACE,
-}, {
-	str:   "debug",
-	level: loggo.DEBUG,
-}, {
-	str:   "DEBUG",
-	level: loggo.DEBUG,
-}, {
-	str:   "info",
-	level: loggo.INFO,
-}, {
-	str:   "INFO",
-	level: loggo.INFO,
-}, {
-	str:   "warn",
-	level: loggo.WARNING,
-}, {
-	str:   "WARN",
-	level: loggo.WARNING,
-}, {
-	str:   "warning",
-	level: loggo.WARNING,
-}, {
-	str:   "WARNING",
-	level: loggo.WARNING,
-}, {
-	str:   "error",
-	level: loggo.ERROR,
-}, {
-	str:   "ERROR",
-	level: loggo.ERROR,
-}, {
-	str:   "critical",
-	level: loggo.CRITICAL,
-}, {
-	str:  "not_specified",
-	fail: true,
-}, {
-	str:  "other",
-	fail: true,
-}, {
-	str:  "",
-	fail: true,
-}}
-
-func (*loggerSuite) TestParseLevel(c *gc.C) {
-	for _, test := range parseLevelTests {
-		level, ok := loggo.ParseLevel(test.str)
-		c.Assert(level, gc.Equals, test.level)
-		c.Assert(ok, gc.Equals, !test.fail)
-	}
-}
-
-var levelStringValueTests = map[loggo.Level]string{
-	loggo.UNSPECIFIED: "UNSPECIFIED",
-	loggo.DEBUG:       "DEBUG",
-	loggo.TRACE:       "TRACE",
-	loggo.INFO:        "INFO",
-	loggo.WARNING:     "WARNING",
-	loggo.ERROR:       "ERROR",
-	loggo.CRITICAL:    "CRITICAL",
-	loggo.Level(42):   "<unknown>", // other values are unknown
-}
-
-func (*loggerSuite) TestLevelStringValue(c *gc.C) {
-	for level, str := range levelStringValueTests {
-		c.Assert(level.String(), gc.Equals, str)
-	}
-}
-
 type stack_error struct {
 	message string
 	stack   []string
